+++ pkg/partition/partition.go
@@ -3,9 +3,9 @@ package partition
 import (
 	"context"
 	"fmt"
-	"github.com/Altinity/clickhouse-backup/pkg/clickhouse"
-	"github.com/Altinity/clickhouse-backup/pkg/common"
-	"github.com/Altinity/clickhouse-backup/pkg/metadata"
+	"github.com/Altinity/clickhouse-backup/v2/pkg/clickhouse"
+	"github.com/Altinity/clickhouse-backup/v2/pkg/common"
+	"github.com/Altinity/clickhouse-backup/v2/pkg/metadata"
 	apexLog "github.com/apex/log"
 	"github.com/google/uuid"
 	"github.com/pkg/errors"