+++ pkg/storage/azblob.go
@@ -6,7 +6,7 @@ import (
 	"crypto/sha256"
 	"encoding/base64"
 	"fmt"
-	"github.com/Altinity/clickhouse-backup/pkg/config"
+	"github.com/Altinity/clickhouse-backup/v2/pkg/config"
 	"github.com/apex/log"
 	"io"
 	"net/url"
@@ -14,7 +14,7 @@ import (
 	"strings"
 	"time"
 
-	x "github.com/Altinity/clickhouse-backup/pkg/storage/azblob"
+	x "github.com/Altinity/clickhouse-backup/v2/pkg/storage/azblob"
 
 	"github.com/Azure/azure-pipeline-go/pipeline"
 	"github.com/Azure/azure-storage-blob-go/azblob"