+++ pkg/storage/object_disk/object_disk.go
@@ -14,9 +14,9 @@ import (
 	"sync"
 	"time"
 
-	"github.com/Altinity/clickhouse-backup/pkg/clickhouse"
-	"github.com/Altinity/clickhouse-backup/pkg/config"
-	"github.com/Altinity/clickhouse-backup/pkg/storage"
+	"github.com/Altinity/clickhouse-backup/v2/pkg/clickhouse"
+	"github.com/Altinity/clickhouse-backup/v2/pkg/config"
+	"github.com/Altinity/clickhouse-backup/v2/pkg/storage"
 	"github.com/antchfx/xmlquery"
 	apexLog "github.com/apex/log"
 	"github.com/puzpuzpuz/xsync"