+++ test/integration/integration_test.go
@@ -5,11 +5,11 @@ package main
 import (
 	"context"
 	"fmt"
-	"github.com/Altinity/clickhouse-backup/pkg/config"
-	"github.com/Altinity/clickhouse-backup/pkg/logcli"
-	"github.com/Altinity/clickhouse-backup/pkg/partition"
-	"github.com/Altinity/clickhouse-backup/pkg/status"
-	"github.com/Altinity/clickhouse-backup/pkg/utils"
+	"github.com/Altinity/clickhouse-backup/v2/pkg/config"
+	"github.com/Altinity/clickhouse-backup/v2/pkg/logcli"
+	"github.com/Altinity/clickhouse-backup/v2/pkg/partition"
+	"github.com/Altinity/clickhouse-backup/v2/pkg/status"
+	"github.com/Altinity/clickhouse-backup/v2/pkg/utils"
 	"github.com/google/uuid"
 	"math/rand"
 	"os"
@@ -22,7 +22,7 @@ import (
 	"testing"
 	"time"
 
-	"github.com/Altinity/clickhouse-backup/pkg/clickhouse"
+	"github.com/Altinity/clickhouse-backup/v2/pkg/clickhouse"
 	"github.com/apex/log"
 	"golang.org/x/mod/semver"